Fixing a number of compiler warnings and other changes needed to compile nicely under Xcode 4
git-svn-id: https://crag.googlecode.com/svn/trunk@307 03cb304c-929a-b8d2-5531-aad0e9782889

+++ src/crag/form/scene/Mesh.cpp
@@ -48,12 +48,12 @@ form::MeshProperties const & form::Mesh::GetProperties() const
 	return properties;
 }
 
-int form::Mesh::GetIndexCount() const
+size_t form::Mesh::GetIndexCount() const
 {
 	return indices.GetSize();
 }
 
-int form::Mesh::GetNumPolys() const
+size_t form::Mesh::GetNumPolys() const
 {
 	return indices.GetSize() / 3;
 }
